Move hasVerificationKey function.
