+++ lib/wallet/src/Cardano/Wallet/Address/Derivation/Shared.hs
@@ -27,7 +27,6 @@ module Cardano.Wallet.Address.Derivation.Shared
     -- * Generation and derivation
     , generateKeyFromSeed
     , unsafeGenerateKeyFromSeed
-    , allCosignerStakingKeys
 
     , purposeCIP1854
     ) where
@@ -36,8 +35,6 @@ import Prelude
 
 import Cardano.Address.Derivation
     ( xpubPublicKey )
-import Cardano.Address.Script
-    ( KeyHash, ScriptTemplate (..) )
 import Cardano.Crypto.Wallet
     ( XPrv, XPub, toXPub, unXPub, xpub )
 import Cardano.Mnemonic
@@ -51,11 +48,9 @@ import Cardano.Wallet.Address.Derivation
     , KeyFingerprint (..)
     , MkKeyFingerprint (..)
     , PersistPublicKey (..)
-    , Role (..)
     , SoftDerivation (..)
     , WalletKey (..)
     , fromHex
-    , hashVerificationKey
     , hex
     , toAddressParts
     )
@@ -90,9 +85,7 @@ import Data.ByteString
 import Data.Proxy
     ( Proxy (..) )
 
-import qualified Cardano.Address.Script as CA
 import qualified Data.ByteString as BS
-import qualified Data.Map.Strict as Map
 
 {-------------------------------------------------------------------------------
                             Sequential Derivation
@@ -132,17 +125,6 @@ instance HardDerivation SharedKey where
 instance SoftDerivation SharedKey where
     deriveAddressPublicKey (SharedKey accXPub) role ix =
         SharedKey $ deriveAddressPublicKeyShelley accXPub role ix
-
-allCosignerStakingKeys
-    :: ScriptTemplate
-    -> [KeyHash]
-allCosignerStakingKeys (ScriptTemplate xpubs _) =
-    map toKeyHash (Map.elems xpubs)
-  where
-    stakingKey accXPub =
-        deriveAddressPublicKey (SharedKey accXPub) MutableAccount minBound
-    toKeyHash =
-        hashVerificationKey CA.Delegation . stakingKey
 
 {-------------------------------------------------------------------------------
                             WalletKey implementation
